Treat `XNN_VALUE_FLAG_EXTERNAL_INPUT | XNN_VALUE_FLAG_EXTERNAL_OUTPUT` as "persistent" tensors
This provides an alternative to the `XNN_VALUE_FLAG_PERSISTENT` approach of implementing persistent tensors. The difference is that while XNNPACK allocates memory for `XNN_VALUE_FLAG_PERSISTENT` tensors, the "inout" tensors are just like any other input/output parameter that the caller passes memory for.

Note that xnnpack.h comments for value flags already say that combinations of `XNN_VALUE_FLAG_EXTERNAL_INPUT` and `XNN_VALUE_FLAG_EXTERNAL_OUTPUT` are supported.

In practice, the *only* actual impact of this change is to work around an assert that requires that values have only one producer (unless the value is persistent).

PiperOrigin-RevId: 750039938