Use /clang: to pass copts to clang-cl
`cl` (not `clang-cl`) works by just allowing the program to use intrinsics, without specifying any target architecture. The `/arch:` flag only controls autovectorization behavior and nothing else.

clang and gcc are different: you must specify architecture flags to use intrinsics for instructions that are only available in that architecture.

clang-cl has a combination of these behaviors that seems broken: it will produce an error if the program uses an intrinsic that requires an architecture flag, like clang/gcc, but it's `/arch:` parameter is lacking the necessary options to actually specify the target architecture. For example, SSE4.1 requires `/arch:AVX`, because there is no `/arch:SSE4.1` flag.

This change uses `/clang:` on clang-cl, to work around this problem.

PiperOrigin-RevId: 843719212